refactor(frontend): extract SchemaForm + share errMessage helper
The inline SchemaForm component inside AdminPluginInstancePage.tsx
(~180 LoC handling string/number/boolean/string[]/secret schema fields,
including ADR-049 sentinel rendering) is now a standalone component at
src/components/form/SchemaForm/ with its own CSS module, exported via
the form/index.ts barrel. SubscriptionsTab and ConfigTab both consume
it via the new import. Identical render shape — strictly an extraction.

errMessage(err, fallback) is hoisted from CredentialsTab into
src/api/fetch.ts (alongside ApiError), and reused from ReauthorizeButton
which had open-coded the same apiErr?.detail ?? apiErr?.message ??
fallback pattern.

The original useFormState + useRedactedSchemaFields proposal was
narrowed once SchemaForm extraction was done: the remaining form-state
duplication carried real variation (1-3 fields, simple-vs-fieldErrors,
ConfigTab's CAS conflict) and a single hook would have been a leaky
abstraction. Sentinel rendering moved into SchemaForm; the only
residual buildPayloadConfig is genuinely page-specific.
