+++ frontend/src/api/fetch.ts
@@ -113,3 +113,12 @@ export async function apiFetchVoid(path: string, init?: RequestInit): Promise<vo
 }
 
 export { baseRequest }
+
+// errMessage extracts a human-readable error message from a TanStack Query
+// onError callback value. TanStack Query types its error as `unknown`; for
+// ApiError it prefers `detail` (server-supplied context) over `message`, and
+// falls back to the caller-supplied default for unknown error shapes.
+export function errMessage(err: unknown, fallback: string): string {
+  const apiErr = err as ApiError
+  return apiErr?.detail ?? apiErr?.message ?? fallback
+}
