+++ frontend/src/components/form/SchemaForm/index.ts
@@ -1,0 +1,2 @@
+export { SchemaForm, REDACTION_SENTINEL } from './SchemaForm'
+export type { SchemaFormProps, SchemaProperty, SchemaShape } from './SchemaForm'
