+++ frontend/src/components/form/index.ts
@@ -2,3 +2,5 @@ export { FieldError } from './FieldError'
 export type { FieldErrorProps } from './FieldError'
 export { ErrorBanner } from './ErrorBanner'
 export type { ErrorBannerProps, BannerIssue } from './ErrorBanner'
+export { SchemaForm, REDACTION_SENTINEL } from './SchemaForm'
+export type { SchemaFormProps, SchemaProperty, SchemaShape } from './SchemaForm'
