+++ frontend/src/pages/AdminPluginInstancePage.module.css
@@ -81,79 +81,6 @@
   line-height: 1.5;
 }
 
-/* Schema form */
-
-.schemaFields {
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-5);
-}
-
-.fieldRow {
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-1);
-}
-
-.fieldLabel {
-  font-size: var(--text-sm);
-  font-weight: var(--weight-medium);
-  color: var(--text-primary);
-}
-
-.fieldHint {
-  font-weight: var(--weight-normal);
-  color: var(--text-muted);
-  font-size: var(--text-xs);
-}
-
-.fieldDesc {
-  margin: 0;
-  font-size: var(--text-xs);
-  color: var(--text-muted);
-}
-
-.checkboxLabel {
-  display: flex;
-  align-items: center;
-  gap: var(--space-2);
-  font-size: var(--text-sm);
-  font-weight: var(--weight-medium);
-  color: var(--text-primary);
-  cursor: pointer;
-}
-
-.input {
-  padding: var(--space-2) var(--space-3);
-  font-size: var(--text-sm);
-  background: var(--bg-elevated);
-  border: 1px solid var(--border-mid);
-  border-radius: 4px;
-  color: var(--text-primary);
-  font-family: var(--font-body);
-  width: 100%;
-  box-sizing: border-box;
-}
-
-.input:focus {
-  outline: 2px solid var(--color-blue);
-  outline-offset: 1px;
-}
-
-.textarea {
-  composes: input;
-  resize: vertical;
-  font-family: var(--font-mono);
-  line-height: 1.5;
-  min-height: 80px;
-}
-
-.noFields {
-  margin: 0;
-  font-size: var(--text-sm);
-  color: var(--text-muted);
-}
-
 /* Save row */
 
 .saveRow {
